DRY out two RTree index maintainers.
