+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/indexes/RTreeIndexHelper.java
@@ -1,0 +1,579 @@
+/*
+ * RTreeIndexHelper.java
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2026 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.apple.foundationdb.record.provider.foundationdb.indexes;
+
+import com.apple.foundationdb.KeyValue;
+import com.apple.foundationdb.ReadTransaction;
+import com.apple.foundationdb.annotation.API;
+import com.apple.foundationdb.async.AsyncIterator;
+import com.apple.foundationdb.async.AsyncUtil;
+import com.apple.foundationdb.async.rtree.ChildSlot;
+import com.apple.foundationdb.async.rtree.ItemSlot;
+import com.apple.foundationdb.async.rtree.Node;
+import com.apple.foundationdb.async.rtree.NodeHelpers;
+import com.apple.foundationdb.async.rtree.OnReadListener;
+import com.apple.foundationdb.async.rtree.OnWriteListener;
+import com.apple.foundationdb.async.rtree.RTree;
+import com.apple.foundationdb.async.rtree.RTreeHilbertCurveHelpers;
+import com.apple.foundationdb.record.CursorStreamingMode;
+import com.apple.foundationdb.record.EndpointType;
+import com.apple.foundationdb.record.ExecuteProperties;
+import com.apple.foundationdb.record.IndexEntry;
+import com.apple.foundationdb.record.PipelineOperation;
+import com.apple.foundationdb.record.RecordCoreException;
+import com.apple.foundationdb.record.RecordCursor;
+import com.apple.foundationdb.record.RecordCursorContinuation;
+import com.apple.foundationdb.record.RecordCursorProto;
+import com.apple.foundationdb.record.RecordCursorResult;
+import com.apple.foundationdb.record.ScanProperties;
+import com.apple.foundationdb.record.TupleRange;
+import com.apple.foundationdb.record.cursors.AsyncIteratorCursor;
+import com.apple.foundationdb.record.cursors.AsyncLockCursor;
+import com.apple.foundationdb.record.cursors.ChainedCursor;
+import com.apple.foundationdb.record.cursors.CursorLimitManager;
+import com.apple.foundationdb.record.cursors.LazyCursor;
+import com.apple.foundationdb.record.locking.LockIdentifier;
+import com.apple.foundationdb.record.provider.common.StoreTimer;
+import com.apple.foundationdb.record.provider.foundationdb.FDBIndexableRecord;
+import com.apple.foundationdb.record.provider.foundationdb.FDBStoreTimer;
+import com.apple.foundationdb.record.provider.foundationdb.IndexMaintainerState;
+import com.apple.foundationdb.record.provider.foundationdb.KeyValueCursor;
+import com.apple.foundationdb.subspace.Subspace;
+import com.apple.foundationdb.tuple.ByteArrayUtil2;
+import com.apple.foundationdb.tuple.Tuple;
+import com.apple.foundationdb.tuple.TupleHelpers;
+import com.google.common.base.Verify;
+import com.google.common.collect.Lists;
+import com.google.protobuf.ByteString;
+import com.google.protobuf.InvalidProtocolBufferException;
+import com.google.protobuf.Message;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.math.BigInteger;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
+import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
+
+/**
+ * Shared maintenance and scan mechanics for index maintainers backed by a Hilbert {@link RTree}. The R-tree stores a
+ * point plus a key suffix per record, partitioned into one tree per <em>prefix</em> tuple (the leading grouping/prefix
+ * columns of the index key). This helper owns everything that does not depend on how the concrete index interprets its
+ * coordinate columns: the prefix skip-scan, the R-tree traversal, continuations, instrumentation listeners, and the
+ * insert/delete path. Concrete maintainers supply the coordinate-specific behavior through {@link PointBuilder} (how a
+ * key's coordinate columns become a stored point) and {@link CoordinateReconstructor} (how stored coordinates are
+ * rendered back into a returned index entry key).
+ *
+ * @see MultidimensionalIndexMaintainer
+ * @see GeospatialRTreeIndexMaintainer
+ */
+@API(API.Status.EXPERIMENTAL)
+public class RTreeIndexHelper {
+    private static final byte NODE_SLOT_INDEX_SUBSPACE_INDICATOR = 0x00;
+
+    private RTreeIndexHelper() {
+    }
+
+    /**
+     * Builds the stored {@link RTree.Point} from the coordinate columns of an index entry key.
+     */
+    @FunctionalInterface
+    public interface PointBuilder {
+        @Nonnull
+        RTree.Point build(@Nonnull List<Object> coordinateItems);
+    }
+
+    /**
+     * Renders the coordinate portion of a returned index entry key from the coordinates stored in the R-tree.
+     */
+    @FunctionalInterface
+    public interface CoordinateReconstructor {
+        @Nonnull
+        List<Object> reconstruct(@Nonnull Tuple storedCoordinates);
+    }
+
+    /**
+     * The subspace holding the R-tree's node-slot secondary index, derived from the index's secondary subspace.
+     * @param state the maintainer state
+     * @return the node-slot index subspace
+     */
+    @Nonnull
+    public static Subspace nodeSlotIndexSubspace(@Nonnull final IndexMaintainerState state) {
+        return state.store.indexSecondarySubspace(state.index).subspace(Tuple.from(NODE_SLOT_INDEX_SUBSPACE_INDICATOR));
+    }
+
+    /**
+     * Scan the R-tree(s) for the given spatial predicate. Iterates each distinct prefix tuple (each a separate R-tree)
+     * within {@code prefixRange}, joining with an R-tree search bounded by {@code mbrPredicate} and {@code suffixRange}
+     * and filtered by {@code containsPredicate}.
+     *
+     * @param state the maintainer state
+     * @param config the R-tree configuration
+     * @param prefixSize number of leading (prefix/grouping) columns partitioning the index into separate R-trees
+     * @param prefixRange range constraining the prefix columns
+     * @param suffixRange range constraining the key suffix, applied both as an R-tree bound and a post-filter
+     * @param mbrPredicate accepts a node's minimum bounding rectangle if the search should descend into it
+     * @param containsPredicate accepts a stored point if it satisfies the (exact) spatial condition
+     * @param coordinateReconstructor renders stored coordinates back into the returned index entry key
+     * @param continuation continuation from a previous scan, or {@code null}
+     * @param scanProperties skip, limit and other scan properties
+     * @return a cursor over the matching index entries
+     */
+    @SuppressWarnings("resource")
+    @Nonnull
+    public static RecordCursor<IndexEntry> scan(@Nonnull final IndexMaintainerState state,
+                                                @Nonnull final RTree.Config config,
+                                                final int prefixSize,
+                                                @Nonnull final TupleRange prefixRange,
+                                                @Nonnull final TupleRange suffixRange,
+                                                @Nonnull final Predicate<RTree.Rectangle> mbrPredicate,
+                                                @Nonnull final Predicate<RTree.Point> containsPredicate,
+                                                @Nonnull final CoordinateReconstructor coordinateReconstructor,
+                                                @Nullable final byte[] continuation,
+                                                @Nonnull final ScanProperties scanProperties) {
+        final ExecuteProperties executeProperties = scanProperties.getExecuteProperties();
+        final ScanProperties innerScanProperties = scanProperties.with(ExecuteProperties::clearSkipAndLimit);
+        final CursorLimitManager cursorLimitManager = new CursorLimitManager(state.context, innerScanProperties);
+        final Subspace indexSubspace = state.indexSubspace;
+        final Subspace nodeSlotIndexSubspace = nodeSlotIndexSubspace(state);
+        final Executor executor = state.context.getExecutor();
+        final FDBStoreTimer timer = Objects.requireNonNull(state.context.getTimer());
+
+        //
+        // Skip-scan through the prefixes in a way that we only consider each distinct prefix. That skip scan
+        // forms the outer of a join with an inner that searches the R-tree for that prefix using the
+        // spatial predicates of the scan bounds.
+        //
+        return RecordCursor.flatMapPipelined(prefixSkipScan(state, prefixSize, timer, prefixRange, innerScanProperties),
+                        (prefixTuple, innerContinuation) -> {
+                            final Subspace rtSubspace;
+                            final Subspace rtNodeSlotIndexSubspace;
+                            if (prefixTuple != null) {
+                                Verify.verify(prefixTuple.size() == prefixSize);
+                                rtSubspace = indexSubspace.subspace(prefixTuple);
+                                rtNodeSlotIndexSubspace = nodeSlotIndexSubspace.subspace(prefixTuple);
+                            } else {
+                                rtSubspace = indexSubspace;
+                                rtNodeSlotIndexSubspace = nodeSlotIndexSubspace;
+                            }
+
+                            final Continuation parsedContinuation = Continuation.fromBytes(innerContinuation);
+                            final BigInteger lastHilbertValue =
+                                    parsedContinuation == null ? null : parsedContinuation.getLastHilbertValue();
+                            final Tuple lastKey = parsedContinuation == null ? null : parsedContinuation.getLastKey();
+
+                            final RTree rTree = new RTree(rtSubspace, rtNodeSlotIndexSubspace, executor, config,
+                                    RTreeHilbertCurveHelpers::hilbertValue, NodeHelpers::newRandomNodeId,
+                                    OnWriteListener.NOOP, new OnRead(cursorLimitManager, timer));
+                            final ReadTransaction transaction = state.context.readTransaction(true);
+                            return new LazyCursor<>(state.context.acquireReadLock(new LockIdentifier(rtSubspace))
+                                    .thenApply(lock -> new AsyncLockCursor<>(lock, new ItemSlotCursor(executor,
+                                            rTree.scan(transaction, lastHilbertValue, lastKey,
+                                                    mbrPredicate,
+                                                    (low, high) -> suffixRange.overlaps(low, high)),
+                                            cursorLimitManager, timer))), executor)
+                                    .filter(itemSlot -> lastHilbertValue == null || lastKey == null ||
+                                                        itemSlot.compareHilbertValueAndKey(lastHilbertValue, lastKey) > 0)
+                                    .filter(itemSlot -> containsPredicate.test(itemSlot.getPosition()))
+                                    .filter(itemSlot -> suffixRange.contains(itemSlot.getKeySuffix()))
+                                    .map(itemSlot -> {
+                                        final List<Object> keyItems = Lists.newArrayList();
+                                        if (prefixTuple != null) {
+                                            keyItems.addAll(prefixTuple.getItems());
+                                        }
+                                        keyItems.addAll(coordinateReconstructor.reconstruct(itemSlot.getPosition().getCoordinates()));
+                                        keyItems.addAll(itemSlot.getKeySuffix().getItems());
+                                        return new IndexEntry(state.index, Tuple.fromList(keyItems), itemSlot.getValue());
+                                    });
+                        },
+                        continuation,
+                        state.store.getPipelineSize(PipelineOperation.INDEX_TO_RECORD))
+                .skipThenLimit(executeProperties.getSkip(), executeProperties.getReturnedRowLimit());
+    }
+
+    @Nonnull
+    private static Function<byte[], RecordCursor<Tuple>> prefixSkipScan(@Nonnull final IndexMaintainerState state,
+                                                                        final int prefixSize,
+                                                                        @Nonnull final StoreTimer timer,
+                                                                        @Nonnull final TupleRange prefixRange,
+                                                                        @Nonnull final ScanProperties innerScanProperties) {
+        final Function<byte[], RecordCursor<Tuple>> outerFunction;
+        if (prefixSize > 0) {
+            outerFunction = outerContinuation -> timer.instrument(MultiDimensionalIndexHelper.Events.MULTIDIMENSIONAL_SKIP_SCAN,
+                    new ChainedCursor<>(state.context,
+                            lastKeyOptional -> nextPrefixTuple(state, prefixRange,
+                                    prefixSize, lastKeyOptional.orElse(null), innerScanProperties),
+                            Tuple::pack,
+                            Tuple::fromBytes,
+                            outerContinuation,
+                            innerScanProperties));
+        } else {
+            outerFunction = outerContinuation -> RecordCursor.fromFuture(CompletableFuture.completedFuture(null));
+        }
+        return outerFunction;
+    }
+
+    @SuppressWarnings({"resource", "PMD.CloseResource"})
+    private static CompletableFuture<Optional<Tuple>> nextPrefixTuple(@Nonnull final IndexMaintainerState state,
+                                                                      @Nonnull final TupleRange prefixRange,
+                                                                      final int prefixSize,
+                                                                      @Nullable final Tuple lastPrefixTuple,
+                                                                      @Nonnull final ScanProperties scanProperties) {
+        final Subspace indexSubspace = state.indexSubspace;
+        final KeyValueCursor cursor;
+        if (lastPrefixTuple == null) {
+            cursor = KeyValueCursor.Builder.withSubspace(indexSubspace)
+                    .setContext(state.context)
+                    .setRange(prefixRange)
+                    .setContinuation(null)
+                    .setScanProperties(scanProperties.setStreamingMode(CursorStreamingMode.ITERATOR)
+                            .with(innerExecuteProperties -> innerExecuteProperties.setReturnedRowLimit(1)))
+                    .build();
+        } else {
+            KeyValueCursor.Builder builder = KeyValueCursor.Builder.withSubspace(indexSubspace)
+                    .setContext(state.context)
+                    .setContinuation(null)
+                    .setScanProperties(scanProperties)
+                    .setScanProperties(scanProperties.setStreamingMode(CursorStreamingMode.ITERATOR)
+                            .with(innerExecuteProperties -> innerExecuteProperties.setReturnedRowLimit(1)));
+
+            cursor = builder.setLow(indexSubspace.pack(lastPrefixTuple), EndpointType.RANGE_EXCLUSIVE)
+                    .setHigh(prefixRange.getHigh(), prefixRange.getHighEndpoint())
+                    .build();
+        }
+
+        return cursor.onNext().thenApply(next -> {
+            cursor.close();
+            if (next.hasNext()) {
+                final KeyValue kv = Objects.requireNonNull(next.get());
+                return Optional.of(TupleHelpers.subTuple(indexSubspace.unpack(kv.getKey()), 0, prefixSize));
+            }
+            return Optional.empty();
+        });
+    }
+
+    /**
+     * Insert or delete the R-tree items for the given index entries.
+     *
+     * @param state the maintainer state
+     * @param config the R-tree configuration
+     * @param prefixSize number of leading (prefix/grouping) columns partitioning the index into separate R-trees
+     * @param dimensionsSize number of coordinate columns following the prefix
+     * @param savedRecord the record being indexed
+     * @param remove {@code true} to delete, {@code false} to insert or update
+     * @param indexEntries the index entries for the record
+     * @param pointBuilder builds the stored point from an entry's coordinate columns
+     * @param <M> the record message type
+     * @return a future that completes when all entries have been maintained
+     */
+    @Nonnull
+    public static <M extends Message> CompletableFuture<Void> update(@Nonnull final IndexMaintainerState state,
+                                                                     @Nonnull final RTree.Config config,
+                                                                     final int prefixSize,
+                                                                     final int dimensionsSize,
+                                                                     @Nonnull final FDBIndexableRecord<M> savedRecord,
+                                                                     final boolean remove,
+                                                                     @Nonnull final List<IndexEntry> indexEntries,
+                                                                     @Nonnull final PointBuilder pointBuilder) {
+        final Subspace indexSubspace = state.indexSubspace;
+        final Subspace nodeSlotIndexSubspace = nodeSlotIndexSubspace(state);
+        final Executor executor = state.context.getExecutor();
+        final FDBStoreTimer timer = Objects.requireNonNull(state.context.getTimer());
+        final var futures = indexEntries.stream().map(indexEntry -> {
+            final var indexKeyItems = indexEntry.getKey().getItems();
+            final Tuple prefixKey = Tuple.fromList(indexKeyItems.subList(0, prefixSize));
+
+            final Subspace rtSubspace;
+            final Subspace rtNodeSlotIndexSubspace;
+            if (prefixSize > 0) {
+                rtSubspace = indexSubspace.subspace(prefixKey);
+                rtNodeSlotIndexSubspace = nodeSlotIndexSubspace.subspace(prefixKey);
+            } else {
+                rtSubspace = indexSubspace;
+                rtNodeSlotIndexSubspace = nodeSlotIndexSubspace;
+            }
+            return state.context.doWithWriteLock(new LockIdentifier(rtSubspace), () -> {
+                final RTree.Point point =
+                        pointBuilder.build(indexKeyItems.subList(prefixSize, prefixSize + dimensionsSize));
+
+                final List<Object> primaryKeyParts = Lists.newArrayList(savedRecord.getPrimaryKey().getItems());
+                state.index.trimPrimaryKey(primaryKeyParts);
+                final List<Object> keySuffixParts =
+                        Lists.newArrayList(indexKeyItems.subList(prefixSize + dimensionsSize, indexKeyItems.size()));
+                keySuffixParts.addAll(primaryKeyParts);
+                final Tuple keySuffix = Tuple.fromList(keySuffixParts);
+                final RTree rTree = new RTree(rtSubspace, rtNodeSlotIndexSubspace, executor, config,
+                        RTreeHilbertCurveHelpers::hilbertValue, NodeHelpers::newRandomNodeId, new OnWrite(timer),
+                        OnReadListener.NOOP);
+                if (remove) {
+                    return rTree.delete(state.transaction, point, keySuffix);
+                } else {
+                    return rTree.insertOrUpdate(state.transaction,
+                            point,
+                            keySuffix,
+                            indexEntry.getValue());
+                }
+            });
+        }).collect(Collectors.toList());
+        return AsyncUtil.whenAll(futures);
+    }
+
+    static class OnRead implements OnReadListener {
+        @Nonnull
+        private final CursorLimitManager cursorLimitManager;
+        @Nonnull
+        private final FDBStoreTimer timer;
+
+        public OnRead(@Nonnull final CursorLimitManager cursorLimitManager,
+                      @Nonnull final FDBStoreTimer timer) {
+            this.cursorLimitManager = cursorLimitManager;
+            this.timer = timer;
+        }
+
+        @Override
+        public <T extends Node> CompletableFuture<T> onAsyncRead(@Nonnull final CompletableFuture<T> future) {
+            return timer.instrument(MultiDimensionalIndexHelper.Events.MULTIDIMENSIONAL_SCAN, future);
+        }
+
+        @Override
+        public void onNodeRead(@Nonnull final Node node) {
+            switch (node.getKind()) {
+                case LEAF:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_LEAF_NODE_READS);
+                    break;
+                case INTERMEDIATE:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_INTERMEDIATE_NODE_READS);
+                    break;
+                default:
+                    throw new RecordCoreException("unsupported kind of node");
+            }
+        }
+
+        @Override
+        public void onKeyValueRead(@Nonnull final Node node, @Nullable final byte[] key, @Nullable final byte[] value) {
+            final int keyLength = key == null ? 0 : key.length;
+            final int valueLength = value == null ? 0 : value.length;
+
+            final int totalLength = keyLength + valueLength;
+            cursorLimitManager.reportScannedBytes(totalLength);
+            cursorLimitManager.tryRecordScan();
+            timer.increment(FDBStoreTimer.Counts.LOAD_INDEX_KEY);
+            timer.increment(FDBStoreTimer.Counts.LOAD_INDEX_KEY_BYTES, keyLength);
+            timer.increment(FDBStoreTimer.Counts.LOAD_INDEX_VALUE_BYTES, valueLength);
+
+            switch (node.getKind()) {
+                case LEAF:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_LEAF_NODE_READ_BYTES, totalLength);
+                    break;
+                case INTERMEDIATE:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_INTERMEDIATE_NODE_READ_BYTES, totalLength);
+                    break;
+                default:
+                    throw new RecordCoreException("unsupported kind of node");
+            }
+        }
+
+        @Override
+        public void onChildNodeDiscard(@Nonnull final ChildSlot childSlot) {
+            timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_CHILD_NODE_DISCARDS);
+        }
+    }
+
+    static class OnWrite implements OnWriteListener {
+        @Nonnull
+        private final FDBStoreTimer timer;
+
+        public OnWrite(@Nonnull final FDBStoreTimer timer) {
+            this.timer = timer;
+        }
+
+        @Override
+        public <T extends Node> CompletableFuture<T> onAsyncReadForWrite(@Nonnull final CompletableFuture<T> future) {
+            return timer.instrument(MultiDimensionalIndexHelper.Events.MULTIDIMENSIONAL_MODIFICATION, future);
+        }
+
+        @Override
+        public void onNodeWritten(@Nonnull final Node node) {
+            switch (node.getKind()) {
+                case LEAF:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_LEAF_NODE_WRITES);
+                    break;
+                case INTERMEDIATE:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_INTERMEDIATE_NODE_WRITES);
+                    break;
+                default:
+                    throw new RecordCoreException("unsupported kind of node");
+            }
+        }
+
+        @Override
+        public void onKeyValueWritten(@Nonnull final Node node, @Nullable final byte[] key, @Nullable final byte[] value) {
+            final int keyLength = key == null ? 0 : key.length;
+            final int valueLength = value == null ? 0 : value.length;
+
+            final int totalLength = keyLength + valueLength;
+            timer.increment(FDBStoreTimer.Counts.SAVE_INDEX_KEY);
+            timer.increment(FDBStoreTimer.Counts.SAVE_INDEX_KEY_BYTES, keyLength);
+            timer.increment(FDBStoreTimer.Counts.SAVE_INDEX_VALUE_BYTES, valueLength);
+
+            switch (node.getKind()) {
+                case LEAF:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_LEAF_NODE_WRITE_BYTES, totalLength);
+                    break;
+                case INTERMEDIATE:
+                    timer.increment(FDBStoreTimer.Counts.MULTIDIMENSIONAL_INTERMEDIATE_NODE_WRITE_BYTES, totalLength);
+                    break;
+                default:
+                    throw new RecordCoreException("unsupported kind of node");
+            }
+        }
+    }
+
+    static class ItemSlotCursor extends AsyncIteratorCursor<ItemSlot> {
+        @Nonnull
+        private final CursorLimitManager cursorLimitManager;
+        @Nonnull
+        private final FDBStoreTimer timer;
+
+        public ItemSlotCursor(@Nonnull final Executor executor, @Nonnull final AsyncIterator<ItemSlot> iterator,
+                              @Nonnull final CursorLimitManager cursorLimitManager, @Nonnull final FDBStoreTimer timer) {
+            super(executor, iterator);
+            this.cursorLimitManager = cursorLimitManager;
+            this.timer = timer;
+        }
+
+        @Nonnull
+        @Override
+        public CompletableFuture<RecordCursorResult<ItemSlot>> onNext() {
+            if (nextResult != null && !nextResult.hasNext()) {
+                // This guard is needed to guarantee that if onNext is called multiple times after the cursor has
+                // returned a result without a value, then the same NoNextReason is returned each time. Without this guard,
+                // one might return SCAN_LIMIT_REACHED (for example) after returning a result with SOURCE_EXHAUSTED because
+                // of the tryRecordScan check.
+                return CompletableFuture.completedFuture(nextResult);
+            } else if (cursorLimitManager.tryRecordScan()) {
+                return iterator.onHasNext().thenApply(hasNext -> {
+                    if (hasNext) {
+                        final ItemSlot itemSlot = iterator.next();
+                        timer.increment(FDBStoreTimer.Counts.LOAD_SCAN_ENTRY);
+                        timer.increment(FDBStoreTimer.Counts.LOAD_KEY_VALUE);
+                        valuesSeen++;
+                        nextResult = RecordCursorResult.withNextValue(itemSlot, new Continuation(itemSlot.getHilbertValue(), itemSlot.getKey()));
+                    } else {
+                        // Source iterator is exhausted.
+                        nextResult = RecordCursorResult.exhausted();
+                    }
+                    return nextResult;
+                });
+            } else { // a limit must have been exceeded
+                final Optional<NoNextReason> stoppedReason = cursorLimitManager.getStoppedReason();
+                if (stoppedReason.isEmpty()) {
+                    throw new RecordCoreException("limit manager stopped cursor but did not report a reason");
+                }
+                Verify.verifyNotNull(nextResult, "should have seen at least one record");
+                nextResult = RecordCursorResult.withoutNextValue(nextResult.getContinuation(), stoppedReason.get());
+                return CompletableFuture.completedFuture(nextResult);
+            }
+        }
+    }
+
+    static class Continuation implements RecordCursorContinuation {
+        @Nullable
+        private final BigInteger lastHilbertValue;
+        @Nullable
+        private final Tuple lastKey;
+        @Nullable
+        private ByteString cachedByteString;
+        @Nullable
+        private byte[] cachedBytes;
+
+        Continuation(@Nullable final BigInteger lastHilbertValue, @Nullable final Tuple lastKey) {
+            this.lastHilbertValue = lastHilbertValue;
+            this.lastKey = lastKey;
+        }
+
+        @Nullable
+        public BigInteger getLastHilbertValue() {
+            return lastHilbertValue;
+        }
+
+        @Nullable
+        public Tuple getLastKey() {
+            return lastKey;
+        }
+
+        @Nonnull
+        @Override
+        public ByteString toByteString() {
+            if (isEnd()) {
+                return ByteString.EMPTY;
+            }
+
+            if (cachedByteString == null) {
+                cachedByteString = RecordCursorProto.MultidimensionalIndexScanContinuation.newBuilder()
+                        .setLastHilbertValue(ByteString.copyFrom(Objects.requireNonNull(lastHilbertValue).toByteArray()))
+                        .setLastKey(ByteString.copyFrom(Objects.requireNonNull(lastKey).pack()))
+                        .build()
+                        .toByteString();
+            }
+            return cachedByteString;
+        }
+
+        @Nullable
+        @Override
+        public byte[] toBytes() {
+            if (isEnd()) {
+                return null;
+            }
+            if (cachedBytes == null) {
+                cachedBytes = toByteString().toByteArray();
+            }
+            return cachedBytes;
+        }
+
+        @Override
+        public boolean isEnd() {
+            return lastHilbertValue == null || lastKey == null;
+        }
+
+        @Nullable
+        static Continuation fromBytes(@Nullable byte[] continuationBytes) {
+            if (continuationBytes != null) {
+                final RecordCursorProto.MultidimensionalIndexScanContinuation parsed;
+                try {
+                    parsed = RecordCursorProto.MultidimensionalIndexScanContinuation.parseFrom(continuationBytes);
+                } catch (InvalidProtocolBufferException ex) {
+                    throw new RecordCoreException("error parsing continuation", ex)
+                            .addLogInfo("raw_bytes", ByteArrayUtil2.loggable(continuationBytes));
+                }
+                return new Continuation(new BigInteger(parsed.getLastHilbertValue().toByteArray()),
+                        Tuple.fromBytes(parsed.getLastKey().toByteArray()));
+            } else {
+                return null;
+            }
+        }
+    }
+}
